refactor: rename BalancesUpdater component file and remove not needed exports
